media: tegra: vi2/4: use 200 msecs as timeout
The syncpt timeouts were specified in jiffies instead of msecs. So
the actual timeout would differ depending on the HZ value.

Use msecs_to_jiffies(200) instead.

bug 2170572

Reviewed-on: https://git-master.nvidia.com/r/1757623
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/media/platform/tegra/camera/vi/mc_common.h
@@ -167,7 +167,7 @@ struct tegra_channel {
 	unsigned int buffer_offset[TEGRA_CSI_BLOCKS];
 	unsigned int buffer_state[QUEUED_BUFFERS];
 	struct vb2_v4l2_buffer *buffers[QUEUED_BUFFERS];
-	unsigned int timeout;
+	unsigned long timeout;
 
 	atomic_t restart_version;
 	int capture_version;
